# produtos/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Pedidos App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /produtos/carrinho/{idProduto}:
    # binds a127 app logic to a route
    x-swagger-router-controller: lista produtos do carrinho de compras 
    delete:
      description: Remove produto do carrinho 
      # used as the method name of the controller
      operationId: remove produto
      parameters:
        - name: idProduto
          in: path
          description: Remove produto do carrinho de compras 
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Produtos"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adiciona produto no carrinho 
      # used as the method name of the controller
      operationId: adicionar produto
      parameters:
        - name: idProduto
          in: path
          description: Adiciona produto no carrinho de compras  
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Produtos"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                     
...
definitions:
  Produtos:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string